+++ Example/Tests/PixelTestCaseTests.swift
@@ -19,7 +19,7 @@ class PixelTestCaseTests: XCTestCase {
         fileManager.fileExists = false
         XCTAssertEqual(fileManager.fileExistsCallCount, 0)
         XCTAssertEqual(fileManager.createDirectoryCallCount, 0)
-        _ = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, option: .dynamicWidthHeight, fileManager: fileManager)
+        _ = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         XCTAssertEqual(fileManager.fileExistsCallCount, 1)
         XCTAssertEqual(fileManager.createDirectoryCallCount, 1)
     }
@@ -30,7 +30,7 @@ class PixelTestCaseTests: XCTestCase {
         fileManager.fileExists = true
         XCTAssertEqual(fileManager.fileExistsCallCount, 0)
         XCTAssertEqual(fileManager.createDirectoryCallCount, 0)
-        _ = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, option: .dynamicWidthHeight, fileManager: fileManager)
+        _ = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         XCTAssertEqual(fileManager.fileExistsCallCount, 1)
         XCTAssertEqual(fileManager.createDirectoryCallCount, 0)
     }
@@ -38,91 +38,91 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_reference_explicit_dynamicWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10), imageType: .reference, option: .dynamicWidthHeight, fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10), imageType: .reference, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_explicit_dynamicWidthHeight_dw_dh@10.0x.png")
     }
     
     func test_fileURL_reference_explicit_dynamicWidth() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10.1), imageType: .reference, option: .dynamicWidth(fixedHeight: 300), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10.1), imageType: .reference, layoutStyle: .dynamicWidth(fixedHeight: 300), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_explicit_dynamicWidth_dw_300.0@10.1x.png")
     }
     
     func test_fileURL_reference_explicit_dynamicHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10.2), imageType: .reference, option: .dynamicHeight(fixedWidth: 300), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10.2), imageType: .reference, layoutStyle: .dynamicHeight(fixedWidth: 300), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_explicit_dynamicHeight_300.0_dh@10.2x.png")
     }
     
     func test_fileURL_reference_explicit_fixedWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10.3), imageType: .reference, option: .fixed(width: 10, height: 20), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(10.3), imageType: .reference, layoutStyle: .fixed(width: 10, height: 20), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_explicit_fixedWidthHeight_10.0_20.0@10.3x.png")
     }
     
     func test_fileURL_diff_explicit_dynamicWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(1), imageType: .diff, option: .dynamicWidthHeight, fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(1), imageType: .diff, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_explicit_dynamicWidthHeight_dw_dh@1.0x.png")
     }
     
     func test_fileURL_diff_explicit_dynamicWidth() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(2), imageType: .diff, option: .dynamicWidth(fixedHeight: 1), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(2), imageType: .diff, layoutStyle: .dynamicWidth(fixedHeight: 1), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_explicit_dynamicWidth_dw_1.0@2.0x.png")
     }
     
     func test_fileURL_diff_explicit_dynamicHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(3), imageType: .diff, option: .dynamicHeight(fixedWidth: 2), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(3), imageType: .diff, layoutStyle: .dynamicHeight(fixedWidth: 2), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_explicit_dynamicHeight_2.0_dh@3.0x.png")
     }
     
     func test_fileURL_diff_explicit_fixedWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(4), imageType: .diff, option: .fixed(width: 3, height: 4), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(4), imageType: .diff, layoutStyle: .fixed(width: 3, height: 4), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_explicit_fixedWidthHeight_3.0_4.0@4.0x.png")
     }
     
     func test_fileURL_failure_explicit_dynamicWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(100), imageType: .failure, option: .dynamicWidthHeight, fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(100), imageType: .failure, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_explicit_dynamicWidthHeight_dw_dh@100.0x.png")
     }
     
     func test_fileURL_failure_explicit_dynamicWidth() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(200), imageType: .failure, option: .dynamicWidth(fixedHeight: 10), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(200), imageType: .failure, layoutStyle: .dynamicWidth(fixedHeight: 10), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_explicit_dynamicWidth_dw_10.0@200.0x.png")
     }
     
     func test_fileURL_failure_explicit_dynamicHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(300), imageType: .failure, option: .dynamicHeight(fixedWidth: 20), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(300), imageType: .failure, layoutStyle: .dynamicHeight(fixedWidth: 20), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_explicit_dynamicHeight_20.0_dh@300.0x.png")
     }
     
     func test_fileURL_failure_explicit_fixedWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(400), imageType: .failure, option: .fixed(width: 30, height: 40), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .explicit(400), imageType: .failure, layoutStyle: .fixed(width: 30, height: 40), fileManager: fileManager)
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_explicit_fixedWidthHeight_30.0_40.0@400.0x.png")
     }
     
     func test_fileURL_reference_native_dynamicWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, option: .dynamicWidthHeight, fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_native_dynamicWidthHeight_dw_dh@\(scale)x.png")
     }
@@ -130,7 +130,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_reference_native_dynamicWidth() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, option: .dynamicWidth(fixedHeight: 300), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, layoutStyle: .dynamicWidth(fixedHeight: 300), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_native_dynamicWidth_dw_300.0@\(scale)x.png")
     }
@@ -138,7 +138,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_reference_native_dynamicHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, option: .dynamicHeight(fixedWidth: 300), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, layoutStyle: .dynamicHeight(fixedWidth: 300), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_native_dynamicHeight_300.0_dh@\(scale)x.png")
     }
@@ -146,7 +146,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_reference_native_fixedWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, option: .fixed(width: 10, height: 20), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .reference, layoutStyle: .fixed(width: 10, height: 20), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Reference/PixelTestCaseTests/fileURL_reference_native_fixedWidthHeight_10.0_20.0@\(scale)x.png")
     }
@@ -154,7 +154,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_diff_native_dynamicWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, option: .dynamicWidthHeight, fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_native_dynamicWidthHeight_dw_dh@\(scale)x.png")
     }
@@ -162,7 +162,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_diff_native_dynamicWidth() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, option: .dynamicWidth(fixedHeight: 1), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, layoutStyle: .dynamicWidth(fixedHeight: 1), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_native_dynamicWidth_dw_1.0@\(scale)x.png")
     }
@@ -170,7 +170,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_diff_native_dynamicHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, option: .dynamicHeight(fixedWidth: 2), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, layoutStyle: .dynamicHeight(fixedWidth: 2), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_native_dynamicHeight_2.0_dh@\(scale)x.png")
     }
@@ -178,7 +178,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_diff_native_fixedWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, option: .fixed(width: 3, height: 4), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .diff, layoutStyle: .fixed(width: 3, height: 4), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Diff/PixelTestCaseTests/fileURL_diff_native_fixedWidthHeight_3.0_4.0@\(scale)x.png")
     }
@@ -186,7 +186,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_failure_native_dynamicWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, option: .dynamicWidthHeight, fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, layoutStyle: .dynamicWidthHeight, fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_native_dynamicWidthHeight_dw_dh@\(scale)x.png")
     }
@@ -194,7 +194,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_failure_native_dynamicWidth() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, option: .dynamicWidth(fixedHeight: 10), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, layoutStyle: .dynamicWidth(fixedHeight: 10), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_native_dynamicWidth_dw_10.0@\(scale)x.png")
     }
@@ -202,7 +202,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_failure_native_dynamicHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, option: .dynamicHeight(fixedWidth: 20), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, layoutStyle: .dynamicHeight(fixedWidth: 20), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_native_dynamicHeight_20.0_dh@\(scale)x.png")
     }
@@ -210,7 +210,7 @@ class PixelTestCaseTests: XCTestCase {
     func test_fileURL_failure_native_fixedWidthHeight() throws {
         let fileManager = MockFileManager()
         let testCase = MockPixelTestCase()
-        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, option: .fixed(width: 30, height: 40), fileManager: fileManager)
+        let fileURL = try testCase.fileURL(forFunction: #function, scale: .native, imageType: .failure, layoutStyle: .fixed(width: 30, height: 40), fileManager: fileManager)
         let scale = UIScreen.main.scale
         XCTAssertEqual(fileURL.relativePath, baseDir + "/PixelTest_TestsSnapshots/Failure/PixelTestCaseTests/fileURL_failure_native_fixedWidthHeight_30.0_40.0@\(scale)x.png")
     }