+++ PixelTest/LayoutStyle.swift
@@ -1,5 +1,5 @@
 //
-//  Option.swift
+//  LayoutStyle.swift
 //  PixelTest
 //
 //  Created by Kane Cheshire on 19/03/2018.
@@ -7,13 +7,13 @@
 
 import Foundation
 
-/// Represents an option for verifying a view.
+/// Represents a layout style for verifying a view.
 ///
 /// - dynamicWidth: The view should have a dynamic width, but fixed height.
 /// - dynamicHeight: The view should have a dynamic height, but fixed width.
 /// - dynamicWidthHeight: The view should have a dynamic width and height.
 /// - fixed: The view should have a fixed width and height.
-public enum Option {
+public enum LayoutStyle {
     case dynamicWidth(fixedHeight: CGFloat)
     case dynamicHeight(fixedWidth: CGFloat)
     case dynamicWidthHeight